fix optimizer

+++ sparkly/index_optimizer/index_optimizer.py
@@ -195,7 +195,7 @@ class IndexOptimizer():
                 .apply(lambda x : x.values)
 
             # convert that to a dataframe
-            res = pd.DataFrame(res.to_dict())
+            res = pd.DataFrame(res.to_dict(), dtype=np.float64)
             # add to current scores
             scores = scores.append(res, ignore_index=True)
             # drop the columns that contain missing values
